Minor refactor of child-grid-row in separate component file.
