+++ projects/igniteui-angular/src/lib/grids/hierarchical-grid/hierarchical-grid.module.ts
@@ -1,12 +1,13 @@
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { IgxGridModule } from '../grid/grid.module';
-import { IgxChildGridRowComponent, IgxHierarchicalGridComponent } from './hierarchical-grid.component';
+import { IgxHierarchicalGridComponent } from './hierarchical-grid.component';
 import { IgxHierarchicalRowComponent } from './hierarchical-row.component';
 import { IgxGridHierarchicalPipe, IgxGridHierarchicalPagingPipe } from './hierarchical-grid.pipes';
 import { IgxRowIslandComponent } from './row-island.component';
 import { IgxHierarchicalGridCellComponent } from './hierarchical-cell.component';
 import { IgxTooltipModule } from '../../directives/tooltip';
 import { ReactiveFormsModule } from '@angular/forms';
+import { IgxChildGridRowComponent } from './child-grid-row.component';
 
 /**
  * @hidden
